consolidate graph
