+++ src/agdb/graph.rs
@@ -1,39 +1,762 @@
-pub mod graph_data;
-pub mod graph_impl;
-pub mod graph_index;
-pub mod graph_storage;
+use crate::collections::vec::DbVec;
+use crate::collections::vec::VecValue;
+use crate::db::db_error::DbError;
+use crate::storage::file_storage::FileStorage;
+use crate::storage::Storage;
+use crate::storage::StorageIndex;
+use crate::utilities::serialize::Serialize;
+use crate::utilities::serialize::SerializeStatic;
+use crate::utilities::stable_hash::StableHash;
+use std::cell::RefCell;
+use std::rc::Rc;
 
-mod graph_data_memory;
-mod graph_data_storage;
-mod graph_data_storage_indexes;
-mod graph_edge;
-mod graph_edge_iterator;
-mod graph_edge_reverse_iterator;
-mod graph_node;
-mod graph_node_iterator;
+pub trait GraphData {
+    fn capacity(&self) -> Result<u64, DbError>;
+    fn commit(&mut self, id: u64) -> Result<(), DbError>;
+    fn free_index(&self) -> Result<i64, DbError>;
+    fn from(&self, index: &GraphIndex) -> Result<i64, DbError>;
+    #[allow(clippy::wrong_self_convention)]
+    fn from_meta(&self, index: &GraphIndex) -> Result<i64, DbError>;
+    fn grow(&mut self) -> Result<(), DbError>;
+    fn node_count(&self) -> Result<u64, DbError>;
+    fn set_from(&mut self, index: &GraphIndex, value: i64) -> Result<(), DbError>;
+    fn set_from_meta(&mut self, index: &GraphIndex, value: i64) -> Result<(), DbError>;
+    fn set_node_count(&mut self, count: u64) -> Result<(), DbError>;
+    fn set_to(&mut self, index: &GraphIndex, value: i64) -> Result<(), DbError>;
+    fn set_to_meta(&mut self, index: &GraphIndex, value: i64) -> Result<(), DbError>;
+    fn to(&self, index: &GraphIndex) -> Result<i64, DbError>;
+    fn to_meta(&self, index: &GraphIndex) -> Result<i64, DbError>;
+    fn transaction(&mut self) -> u64;
+}
 
-use self::graph_data_memory::GraphDataMemory;
-use self::graph_impl::GraphImpl;
+pub(crate) struct GraphDataStorageIndexes {
+    pub(crate) from: StorageIndex,
+    pub(crate) to: StorageIndex,
+    pub(crate) from_meta: StorageIndex,
+    pub(crate) to_meta: StorageIndex,
+}
 
-pub type Graph = GraphImpl<GraphDataMemory>;
+impl Serialize for GraphDataStorageIndexes {
+    fn deserialize(bytes: &[u8]) -> Result<Self, DbError> {
+        Ok(GraphDataStorageIndexes {
+            from: StorageIndex::deserialize(bytes)?,
+            to: StorageIndex::deserialize(
+                &bytes[(StorageIndex::serialized_size_static() as usize)..],
+            )?,
+            from_meta: StorageIndex::deserialize(
+                &bytes[(StorageIndex::serialized_size_static() as usize * 2)..],
+            )?,
+            to_meta: StorageIndex::deserialize(
+                &bytes[(StorageIndex::serialized_size_static() as usize * 3)..],
+            )?,
+        })
+    }
 
-impl Graph {
-    pub fn new() -> Graph {
-        Graph {
-            data: GraphDataMemory::default(),
+    fn serialize(&self) -> Vec<u8> {
+        let mut bytes: Vec<u8> = vec![];
+        bytes.reserve(4 * StorageIndex::serialized_size_static() as usize);
+
+        bytes.extend(self.from.serialize());
+        bytes.extend(self.to.serialize());
+        bytes.extend(self.from_meta.serialize());
+        bytes.extend(self.to_meta.serialize());
+
+        bytes
+    }
+
+    fn serialized_size(&self) -> u64 {
+        Self::serialized_size_static()
+    }
+}
+
+impl SerializeStatic for GraphDataStorageIndexes {}
+
+pub struct GraphDataStorage<S = FileStorage>
+where
+    S: Storage,
+{
+    pub(crate) storage: Rc<RefCell<S>>,
+    pub(crate) storage_index: StorageIndex,
+    pub(crate) from: DbVec<i64, S>,
+    pub(crate) to: DbVec<i64, S>,
+    pub(crate) from_meta: DbVec<i64, S>,
+    pub(crate) to_meta: DbVec<i64, S>,
+}
+
+impl<S> GraphDataStorage<S>
+where
+    S: Storage,
+{
+    pub fn new(storage: Rc<RefCell<S>>) -> Result<Self, DbError> {
+        let id = storage.borrow_mut().transaction();
+
+        let mut from = DbVec::<i64, S>::new(storage.clone())?;
+        from.push(&0)?;
+        let mut to = DbVec::<i64, S>::new(storage.clone())?;
+        to.push(&0)?;
+        let mut from_meta = DbVec::<i64, S>::new(storage.clone())?;
+        from_meta.push(&i64::MIN)?;
+        let mut to_meta = DbVec::<i64, S>::new(storage.clone())?;
+        to_meta.push(&0)?;
+
+        let indexes = GraphDataStorageIndexes {
+            from: from.storage_index(),
+            to: to.storage_index(),
+            from_meta: from_meta.storage_index(),
+            to_meta: to_meta.storage_index(),
+        };
+
+        let index = storage.borrow_mut().insert(&indexes)?;
+
+        storage.borrow_mut().commit(id)?;
+
+        Ok(GraphDataStorage::<S> {
+            storage,
+            storage_index: index,
+            from,
+            to,
+            from_meta,
+            to_meta,
+        })
+    }
+
+    pub fn from_storage(
+        storage: Rc<RefCell<S>>,
+        storage_index: StorageIndex,
+    ) -> Result<Self, DbError> {
+        let indexes = storage
+            .borrow_mut()
+            .value::<GraphDataStorageIndexes>(storage_index)?;
+
+        let from = DbVec::<i64, S>::from_storage(storage.clone(), indexes.from)?;
+        let to = DbVec::<i64, S>::from_storage(storage.clone(), indexes.to)?;
+        let from_meta = DbVec::<i64, S>::from_storage(storage.clone(), indexes.from_meta)?;
+        let to_meta = DbVec::<i64, S>::from_storage(storage.clone(), indexes.to_meta)?;
+
+        Ok(GraphDataStorage::<S> {
+            storage,
+            storage_index,
+            from,
+            to,
+            from_meta,
+            to_meta,
+        })
+    }
+}
+
+impl<Data> GraphData for GraphDataStorage<Data>
+where
+    Data: Storage,
+{
+    fn capacity(&self) -> Result<u64, DbError> {
+        Ok(self.from.len())
+    }
+
+    fn commit(&mut self, id: u64) -> Result<(), DbError> {
+        self.storage.borrow_mut().commit(id)
+    }
+
+    fn free_index(&self) -> Result<i64, DbError> {
+        self.from_meta.value(0)
+    }
+
+    fn from(&self, index: &GraphIndex) -> Result<i64, DbError> {
+        self.from.value(index.as_u64())
+    }
+
+    fn from_meta(&self, index: &GraphIndex) -> Result<i64, DbError> {
+        self.from_meta.value(index.as_u64())
+    }
+
+    fn grow(&mut self) -> Result<(), DbError> {
+        self.from.push(&0)?;
+        self.to.push(&0)?;
+        self.from_meta.push(&0)?;
+        self.to_meta.push(&0)
+    }
+
+    fn node_count(&self) -> Result<u64, DbError> {
+        Ok(self.to_meta.value(0)? as u64)
+    }
+
+    fn set_from(&mut self, index: &GraphIndex, value: i64) -> Result<(), DbError> {
+        self.from.replace(index.as_u64(), &value)?;
+        Ok(())
+    }
+
+    fn set_from_meta(&mut self, index: &GraphIndex, value: i64) -> Result<(), DbError> {
+        self.from_meta.replace(index.as_u64(), &value)?;
+        Ok(())
+    }
+
+    fn set_node_count(&mut self, count: u64) -> Result<(), DbError> {
+        self.to_meta.replace(0, &(count as i64))?;
+        Ok(())
+    }
+
+    fn set_to(&mut self, index: &GraphIndex, value: i64) -> Result<(), DbError> {
+        self.to.replace(index.as_u64(), &value)?;
+        Ok(())
+    }
+
+    fn set_to_meta(&mut self, index: &GraphIndex, value: i64) -> Result<(), DbError> {
+        self.to_meta.replace(index.as_u64(), &value)?;
+        Ok(())
+    }
+
+    fn to(&self, index: &GraphIndex) -> Result<i64, DbError> {
+        self.to.value(index.as_u64())
+    }
+
+    fn to_meta(&self, index: &GraphIndex) -> Result<i64, DbError> {
+        self.to_meta.value(index.as_u64())
+    }
+
+    fn transaction(&mut self) -> u64 {
+        self.storage.borrow_mut().transaction()
+    }
+}
+
+#[derive(Clone, Copy, Debug, Default, Eq, Ord, Hash, PartialEq, PartialOrd)]
+pub struct GraphIndex {
+    pub index: i64,
+}
+
+impl GraphIndex {
+    pub fn is_edge(&self) -> bool {
+        self.index < 0
+    }
+
+    pub fn is_node(&self) -> bool {
+        0 < self.index
+    }
+
+    pub fn is_valid(&self) -> bool {
+        self.index != 0
+    }
+
+    pub(crate) fn as_u64(&self) -> u64 {
+        if self.is_edge() {
+            (-self.index) as u64
+        } else {
+            self.index as u64
         }
+    }
+}
+
+impl From<i64> for GraphIndex {
+    fn from(index: i64) -> Self {
+        Self { index }
+    }
+}
+
+impl StableHash for GraphIndex {
+    fn stable_hash(&self) -> u64 {
+        self.index.stable_hash()
+    }
+}
+
+impl Serialize for GraphIndex {
+    fn serialize(&self) -> Vec<u8> {
+        self.index.serialize()
+    }
+
+    fn deserialize(bytes: &[u8]) -> Result<Self, DbError> {
+        Ok(Self {
+            index: i64::deserialize(bytes)?,
+        })
+    }
+
+    fn serialized_size(&self) -> u64 {
+        self.index.serialized_size()
+    }
+}
+
+impl SerializeStatic for GraphIndex {}
+
+impl VecValue for GraphIndex {
+    fn storage_len() -> u64 {
+        Self::serialized_size_static()
+    }
+}
+
+pub struct GraphNode<'a, Data>
+where
+    Data: GraphData,
+{
+    pub(crate) graph: &'a GraphImpl<Data>,
+    pub(crate) index: GraphIndex,
+}
+
+impl<'a, Data> GraphNode<'a, Data>
+where
+    Data: GraphData,
+{
+    #[allow(dead_code)]
+    pub fn index(&self) -> GraphIndex {
+        self.index
+    }
+
+    pub fn edge_iter_from(&self) -> GraphEdgeIterator<Data> {
+        GraphEdgeIterator {
+            graph: self.graph,
+            index: self.graph.first_edge_from(&self.index).unwrap_or_default(),
+        }
+    }
+
+    pub fn edge_iter_to(&self) -> GraphEdgeReverseIterator<Data> {
+        GraphEdgeReverseIterator {
+            graph: self.graph,
+            index: self.graph.first_edge_to(&self.index).unwrap_or_default(),
+        }
+    }
+}
+
+pub struct GraphNodeIterator<'a, Data>
+where
+    Data: GraphData,
+{
+    pub(crate) graph: &'a GraphImpl<Data>,
+    pub(crate) index: GraphIndex,
+}
+
+impl<'a, Data> Iterator for GraphNodeIterator<'a, Data>
+where
+    Data: GraphData,
+{
+    type Item = GraphNode<'a, Data>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if let Some(next) = self.graph.next_node(&self.index).unwrap_or(None) {
+            self.index = next;
+            return Some(GraphNode {
+                graph: self.graph,
+                index: self.index,
+            });
+        }
+
+        None
+    }
+}
+
+pub struct GraphEdge<'a, Data>
+where
+    Data: GraphData,
+{
+    pub(crate) graph: &'a GraphImpl<Data>,
+    pub(crate) index: GraphIndex,
+}
+
+impl<'a, Data> GraphEdge<'a, Data>
+where
+    Data: GraphData,
+{
+    pub fn index(&self) -> GraphIndex {
+        self.index
+    }
+
+    pub fn index_from(&self) -> GraphIndex {
+        self.graph.edge_from(&self.index)
+    }
+
+    pub fn index_to(&self) -> GraphIndex {
+        self.graph.edge_to(&self.index)
+    }
+}
+
+pub struct GraphEdgeIterator<'a, Data>
+where
+    Data: GraphData,
+{
+    pub(crate) graph: &'a GraphImpl<Data>,
+    pub(crate) index: GraphIndex,
+}
+
+impl<'a, Data> Iterator for GraphEdgeIterator<'a, Data>
+where
+    Data: GraphData,
+{
+    type Item = GraphEdge<'a, Data>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if !self.index.is_valid() {
+            return None;
+        }
+
+        let current_index = self.index;
+
+        self.index = self.graph.next_edge_from(&self.index).unwrap_or_default();
+
+        Some(GraphEdge {
+            graph: self.graph,
+            index: current_index,
+        })
+    }
+}
+
+pub struct GraphEdgeReverseIterator<'a, Data>
+where
+    Data: GraphData,
+{
+    pub(crate) graph: &'a GraphImpl<Data>,
+    pub(crate) index: GraphIndex,
+}
+
+impl<'a, Data> Iterator for GraphEdgeReverseIterator<'a, Data>
+where
+    Data: GraphData,
+{
+    type Item = GraphEdge<'a, Data>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if !self.index.is_valid() {
+            return None;
+        }
+
+        let current_index = self.index;
+
+        self.index = self.graph.next_edge_to(&self.index).unwrap_or_default();
+
+        Some(GraphEdge {
+            graph: self.graph,
+            index: current_index,
+        })
+    }
+}
+
+pub struct GraphImpl<Data>
+where
+    Data: GraphData,
+{
+    pub(crate) data: Data,
+}
+
+impl<Data> GraphImpl<Data>
+where
+    Data: GraphData,
+{
+    pub fn edge(&self, index: &GraphIndex) -> Option<GraphEdge<Data>> {
+        if self.validate_edge(index).is_err() {
+            return None;
+        }
+
+        Some(GraphEdge {
+            graph: self,
+            index: *index,
+        })
+    }
+
+    #[allow(dead_code)]
+    pub fn node_count(&self) -> Result<u64, DbError> {
+        self.data.node_count()
+    }
+
+    pub fn insert_edge(
+        &mut self,
+        from: &GraphIndex,
+        to: &GraphIndex,
+    ) -> Result<GraphIndex, DbError> {
+        self.validate_node(from)?;
+        self.validate_node(to)?;
+
+        let id = self.data.transaction();
+        let index = GraphIndex::from(-self.get_free_index()?);
+        self.set_edge(&index, from, to)?;
+        self.data.commit(id)?;
+
+        Ok(index)
+    }
+
+    pub fn insert_node(&mut self) -> Result<GraphIndex, DbError> {
+        let id = self.data.transaction();
+        let index = GraphIndex::from(self.get_free_index()?);
+        let count = self.data.node_count()?;
+        self.data.set_node_count(count + 1)?;
+        self.data.commit(id)?;
+
+        Ok(index)
+    }
+
+    pub fn node(&self, index: &GraphIndex) -> Option<GraphNode<Data>> {
+        if self.validate_node(index).is_err() {
+            return None;
+        }
+
+        Some(GraphNode {
+            graph: self,
+            index: *index,
+        })
+    }
+
+    #[allow(dead_code)]
+    pub fn node_iter(&self) -> GraphNodeIterator<Data> {
+        GraphNodeIterator {
+            graph: self,
+            index: GraphIndex::default(),
+        }
+    }
+
+    pub fn remove_edge(&mut self, index: &GraphIndex) -> Result<(), DbError> {
+        if self.validate_edge(index).is_err() {
+            return Ok(());
+        }
+
+        let id = self.data.transaction();
+        self.remove_from_edge(index)?;
+        self.remove_to_edge(index)?;
+        self.free_index(&GraphIndex::from(-index.index))?;
+
+        self.data.commit(id)
+    }
+
+    pub fn remove_node(&mut self, index: &GraphIndex) -> Result<(), DbError> {
+        if self.validate_node(index).is_err() {
+            return Ok(());
+        }
+
+        let id = self.data.transaction();
+        self.remove_from_edges(index)?;
+        self.remove_to_edges(index)?;
+        self.free_index(index)?;
+
+        let count = self.data.node_count()?;
+        self.data.set_node_count(count - 1)?;
+
+        self.data.commit(id)
+    }
+
+    pub fn first_edge_from(&self, index: &GraphIndex) -> Result<GraphIndex, DbError> {
+        Ok(GraphIndex::from(-self.data.from(index)?))
+    }
+
+    pub fn first_edge_to(&self, index: &GraphIndex) -> Result<GraphIndex, DbError> {
+        Ok(GraphIndex::from(-self.data.to(index)?))
+    }
+
+    pub(crate) fn edge_from(&self, index: &GraphIndex) -> GraphIndex {
+        GraphIndex::from(-self.data.from(index).unwrap_or_default())
+    }
+
+    pub(crate) fn edge_to(&self, index: &GraphIndex) -> GraphIndex {
+        GraphIndex::from(-self.data.to(index).unwrap_or_default())
+    }
+
+    fn free_index(&mut self, index: &GraphIndex) -> Result<(), DbError> {
+        let next_free = self.data.from_meta(&GraphIndex::default())?;
+        self.data.set_from_meta(index, next_free)?;
+        self.data
+            .set_from_meta(&GraphIndex::default(), -index.index)
+    }
+
+    fn get_free_index(&mut self) -> Result<i64, DbError> {
+        let mut index = self.data.free_index()?;
+
+        if index == i64::MIN {
+            index = self.data.capacity()? as i64;
+            self.data.grow()?;
+
+            Ok(index)
+        } else {
+            let next = self.data.from_meta(&GraphIndex::from(-index))?;
+            self.data.set_from_meta(&GraphIndex::default(), next)?;
+            self.data.set_from_meta(&GraphIndex::from(-index), 0)?;
+
+            Ok(-index)
+        }
+    }
+
+    fn invalid_index(index: &GraphIndex) -> DbError {
+        DbError::from(format!("'{}' is invalid index", index.index))
+    }
+
+    fn is_removed_index(&self, index: &GraphIndex) -> Result<bool, DbError> {
+        Ok(self.data.from_meta(index)? < 0)
+    }
+
+    fn is_valid_edge(&self, index: &GraphIndex) -> Result<bool, DbError> {
+        Ok(self.data.from(index)? < 0)
+    }
+
+    fn is_valid_index(&self, index: &GraphIndex) -> Result<bool, DbError> {
+        Ok(index.is_valid()
+            && index.as_u64() < self.data.capacity()?
+            && !self.is_removed_index(index)?)
+    }
+
+    fn is_valid_node(&self, index: &GraphIndex) -> Result<bool, DbError> {
+        Ok(0 <= self.data.from(index)?)
+    }
+
+    pub(crate) fn next_edge_from(&self, index: &GraphIndex) -> Result<GraphIndex, DbError> {
+        Ok(GraphIndex::from(-self.data.from_meta(index)?))
+    }
+
+    pub(crate) fn next_edge_to(&self, index: &GraphIndex) -> Result<GraphIndex, DbError> {
+        Ok(GraphIndex::from(-self.data.to_meta(index)?))
+    }
+
+    pub(crate) fn next_node(&self, index: &GraphIndex) -> Result<Option<GraphIndex>, DbError> {
+        for i in (index.index + 1)..(self.data.capacity()? as i64) {
+            let next = GraphIndex::from(i);
+            if self.is_valid_node(&next)? && !self.is_removed_index(&next)? {
+                return Ok(Some(next));
+            }
+        }
+
+        Ok(None)
+    }
+
+    fn remove_from_edge(&mut self, index: &GraphIndex) -> Result<(), DbError> {
+        let node = GraphIndex::from(-self.data.from(index)?);
+        let first = GraphIndex::from(-self.data.from(&node)?);
+        let next = self.data.from_meta(index)?;
+
+        if first == *index {
+            self.data.set_from(&node, next)?;
+        } else {
+            let mut previous = first;
+
+            while self.data.from_meta(&previous)? != -index.index {
+                previous = GraphIndex::from(self.data.from_meta(&previous)?);
+            }
+
+            self.data.set_from_meta(&previous, next)?;
+        }
+
+        let count = self.data.from_meta(&node)?;
+        self.data.set_from_meta(&node, count - 1)
+    }
+
+    fn remove_from_edges(&mut self, index: &GraphIndex) -> Result<(), DbError> {
+        let mut edge = GraphIndex::from(-self.data.from(index)?);
+
+        while edge.is_valid() {
+            self.remove_to_edge(&edge)?;
+            let current_index = -edge.index;
+            edge = GraphIndex::from(-self.data.from_meta(&edge)?);
+            self.free_index(&GraphIndex::from(current_index))?;
+        }
+
+        Ok(())
+    }
+
+    fn remove_to_edge(&mut self, index: &GraphIndex) -> Result<(), DbError> {
+        let node = GraphIndex::from(-self.data.to(index)?);
+        let first = GraphIndex::from(-self.data.to(&node)?);
+        let next = self.data.to_meta(index)?;
+
+        if first == *index {
+            self.data.set_to(&node, next)?;
+        } else {
+            let mut previous = first;
+
+            while self.data.to_meta(&previous)? != -index.index {
+                previous = GraphIndex::from(self.data.to_meta(&previous)?);
+            }
+
+            self.data.set_to_meta(&previous, next)?;
+        }
+
+        let count = self.data.to_meta(&node)?;
+        self.data.set_to_meta(&node, count - 1)
+    }
+
+    fn remove_to_edges(&mut self, index: &GraphIndex) -> Result<(), DbError> {
+        let mut edge = GraphIndex::from(-self.data.to(index)?);
+
+        while edge.is_valid() {
+            self.remove_from_edge(&edge)?;
+            let current_index = -edge.index;
+            edge = GraphIndex::from(-self.data.to_meta(&edge)?);
+            self.free_index(&GraphIndex::from(current_index))?;
+        }
+
+        Ok(())
+    }
+
+    fn set_edge(
+        &mut self,
+        index: &GraphIndex,
+        from: &GraphIndex,
+        to: &GraphIndex,
+    ) -> Result<(), DbError> {
+        self.data.set_from(index, -from.index)?;
+        self.data.set_to(index, -to.index)?;
+        self.update_from_edge(from, index)?;
+        self.update_to_edge(to, index)
+    }
+
+    fn update_from_edge(&mut self, node: &GraphIndex, edge: &GraphIndex) -> Result<(), DbError> {
+        let next = self.data.from(node)?;
+        self.data.set_from_meta(edge, next)?;
+        self.data.set_from(node, -edge.index)?;
+
+        let count = self.data.from_meta(node)?;
+        self.data.set_from_meta(node, count + 1)
+    }
+
+    fn update_to_edge(&mut self, node: &GraphIndex, edge: &GraphIndex) -> Result<(), DbError> {
+        let next = self.data.to(node)?;
+        self.data.set_to_meta(edge, next)?;
+        self.data.set_to(node, -edge.index)?;
+
+        let count = self.data.to_meta(node)?;
+        self.data.set_to_meta(node, count + 1)
+    }
+
+    fn validate_edge(&self, index: &GraphIndex) -> Result<(), DbError> {
+        if !self.is_valid_index(index)? || !self.is_valid_edge(index)? {
+            return Err(Self::invalid_index(index));
+        }
+
+        Ok(())
+    }
+
+    fn validate_node(&self, index: &GraphIndex) -> Result<(), DbError> {
+        if !self.is_valid_index(index)? || !self.is_valid_node(index)? {
+            return Err(Self::invalid_index(index));
+        }
+
+        Ok(())
+    }
+}
+
+pub type DbGraph<Data = FileStorage> = GraphImpl<GraphDataStorage<Data>>;
+
+impl<Data> DbGraph<Data>
+where
+    Data: Storage,
+{
+    pub fn storage_index(&self) -> StorageIndex {
+        self.data.storage_index
+    }
+
+    pub fn new(storage: Rc<RefCell<Data>>) -> Result<Self, DbError> {
+        Ok(DbGraph {
+            data: GraphDataStorage::<Data>::new(storage)?,
+        })
+    }
+
+    pub fn from_storage(storage: Rc<RefCell<Data>>, index: StorageIndex) -> Result<Self, DbError> {
+        Ok(DbGraph {
+            data: GraphDataStorage::<Data>::from_storage(storage, index)?,
+        })
     }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::db::db_error::DbError;
-    use crate::graph::graph_index::GraphIndex;
+    use crate::test_utilities::test_file::TestFile;
 
     #[test]
     fn edge_from_index() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
+
         let from = graph.insert_node().unwrap();
         let to = graph.insert_node().unwrap();
         let index = graph.insert_edge(&from, &to).unwrap();
@@ -43,14 +766,22 @@ mod tests {
 
     #[test]
     fn edge_from_index_missing() {
-        let graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let graph = DbGraph::new(storage).unwrap();
 
         assert!(graph.edge(&GraphIndex::from(-3)).is_none());
     }
 
     #[test]
     fn edge_iteration() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
         let node1 = graph.insert_node().unwrap();
         let node2 = graph.insert_node().unwrap();
 
@@ -68,49 +799,12 @@ mod tests {
     }
 
     #[test]
-    fn edge_iteration_reverse() {
-        let mut graph = Graph::new();
-        let node1 = graph.insert_node().unwrap();
-        let node2 = graph.insert_node().unwrap();
-
-        let edge1 = graph.insert_edge(&node1, &node2).unwrap();
-        let edge2 = graph.insert_edge(&node1, &node2).unwrap();
-        let edge3 = graph.insert_edge(&node1, &node2).unwrap();
-
-        let mut actual = Vec::<GraphIndex>::new();
-
-        for edge in graph.node(&node2).unwrap().edge_iter_to() {
-            actual.push(edge.index());
-        }
-
-        assert_eq!(actual, vec![edge3, edge2, edge1]);
-    }
-
-    #[test]
-    fn index_from() {
-        let mut graph = Graph::new();
-        let node1 = graph.insert_node().unwrap();
-        let node2 = graph.insert_node().unwrap();
-
-        let index = graph.insert_edge(&node1, &node2).unwrap();
-
-        assert_eq!(graph.edge(&index).unwrap().index_from(), node1);
-    }
-
-    #[test]
-    fn index_to() {
-        let mut graph = Graph::new();
-        let node1 = graph.insert_node().unwrap();
-        let node2 = graph.insert_node().unwrap();
-
-        let index = graph.insert_edge(&node1, &node2).unwrap();
-
-        assert_eq!(graph.edge(&index).unwrap().index_to(), node2);
-    }
-
-    #[test]
     fn insert_edge() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
         let from = graph.insert_node().unwrap();
         let to = graph.insert_node().unwrap();
 
@@ -119,7 +813,11 @@ mod tests {
 
     #[test]
     fn insert_edge_after_removed() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
         let from = graph.insert_node().unwrap();
         let to = graph.insert_node().unwrap();
         let index = graph.insert_edge(&from, &to).unwrap();
@@ -131,7 +829,11 @@ mod tests {
 
     #[test]
     fn insert_edge_after_several_removed() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
         let from = graph.insert_node().unwrap();
         let to = graph.insert_node().unwrap();
         let index1 = graph.insert_edge(&from, &to).unwrap();
@@ -146,7 +848,11 @@ mod tests {
 
     #[test]
     fn insert_edge_invalid_from() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
 
         assert_eq!(
             graph.insert_edge(&GraphIndex::from(1), &GraphIndex::from(2)),
@@ -156,7 +862,11 @@ mod tests {
 
     #[test]
     fn insert_edge_invalid_to() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
         let from = graph.insert_node().unwrap();
 
         assert_eq!(
@@ -167,26 +877,38 @@ mod tests {
 
     #[test]
     fn insert_node() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
 
         assert_eq!(graph.insert_node(), Ok(GraphIndex::from(1)));
     }
 
     #[test]
     fn insert_node_after_removal() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
         graph.insert_node().unwrap();
         let index = graph.insert_node().unwrap();
         graph.insert_node().unwrap();
 
         graph.remove_node(&index).unwrap();
 
-        assert_eq!(graph.insert_node(), Ok(index));
+        assert_eq!(graph.insert_node().unwrap(), index);
     }
 
     #[test]
     fn node_count() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
 
         assert_eq!(graph.node_count().unwrap(), 0);
 
@@ -194,16 +916,20 @@ mod tests {
         let index = graph.insert_node().unwrap();
         graph.insert_node().unwrap();
 
-        assert_eq!(graph.node_count(), Ok(3));
+        assert_eq!(graph.node_count().unwrap(), 3);
 
         graph.remove_node(&index).unwrap();
 
-        assert_eq!(graph.node_count(), Ok(2));
+        assert_eq!(graph.node_count().unwrap(), 2);
     }
 
     #[test]
     fn node_from_index() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
         let index = graph.insert_node().unwrap();
 
         assert_eq!(graph.node(&index).unwrap().index(), index);
@@ -211,7 +937,11 @@ mod tests {
 
     #[test]
     fn node_from_index_missing() {
-        let graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let graph = DbGraph::new(storage).unwrap();
 
         let node = graph.node(&GraphIndex::from(1));
 
@@ -220,7 +950,11 @@ mod tests {
 
     #[test]
     fn node_iteration() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
         let node1 = graph.insert_node().unwrap();
         let node2 = graph.insert_node().unwrap();
         let node3 = graph.insert_node().unwrap();
@@ -237,7 +971,11 @@ mod tests {
 
     #[test]
     fn node_iteration_with_removed_nodes() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
         let node1 = graph.insert_node().unwrap();
         let node2 = graph.insert_node().unwrap();
         let node3 = graph.insert_node().unwrap();
@@ -259,7 +997,11 @@ mod tests {
 
     #[test]
     fn remove_edge_circular() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
         let node = graph.insert_node().unwrap();
         let index = graph.insert_edge(&node, &node).unwrap();
 
@@ -270,7 +1012,11 @@ mod tests {
 
     #[test]
     fn remove_edge_first() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
         let from = graph.insert_node().unwrap();
         let to = graph.insert_node().unwrap();
         let index1 = graph.insert_edge(&from, &to).unwrap();
@@ -286,7 +1032,11 @@ mod tests {
 
     #[test]
     fn remove_edge_last() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
         let from = graph.insert_node().unwrap();
         let to = graph.insert_node().unwrap();
         let index1 = graph.insert_edge(&from, &to).unwrap();
@@ -302,7 +1052,11 @@ mod tests {
 
     #[test]
     fn remove_edge_middle() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
         let from = graph.insert_node().unwrap();
         let to = graph.insert_node().unwrap();
         let index1 = graph.insert_edge(&from, &to).unwrap();
@@ -318,13 +1072,21 @@ mod tests {
 
     #[test]
     fn remove_edge_missing() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
         graph.remove_edge(&GraphIndex::from(-3)).unwrap();
     }
 
     #[test]
     fn remove_edge_only() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
         let from = graph.insert_node().unwrap();
         let to = graph.insert_node().unwrap();
         let index = graph.insert_edge(&from, &to).unwrap();
@@ -336,7 +1098,11 @@ mod tests {
 
     #[test]
     fn remove_node_circular_edge() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
         let index = graph.insert_node().unwrap();
         let edge = graph.insert_edge(&index, &index).unwrap();
 
@@ -348,7 +1114,11 @@ mod tests {
 
     #[test]
     fn remove_node_only() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
         let index = graph.insert_node().unwrap();
 
         graph.remove_node(&index).unwrap();
@@ -358,13 +1128,21 @@ mod tests {
 
     #[test]
     fn remove_node_missing() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
         graph.remove_node(&GraphIndex::from(1)).unwrap();
     }
 
     #[test]
     fn remove_nodes_with_edges() {
-        let mut graph = Graph::new();
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+        let mut graph = DbGraph::new(storage).unwrap();
 
         let node1 = graph.insert_node().unwrap();
         let node2 = graph.insert_node().unwrap();
@@ -395,13 +1173,43 @@ mod tests {
     }
 
     #[test]
-    fn reinsert_node_with_edge() {
-        let mut graph = Graph::new();
-        let index = graph.insert_node().unwrap();
-        graph.insert_edge(&index, &index).unwrap();
-        graph.remove_node(&index).unwrap();
-        let new_index = graph.insert_node().unwrap();
-        assert_eq!(index, new_index);
-        graph.insert_edge(&new_index, &new_index).unwrap();
+    fn restore_from_file() {
+        let test_file = TestFile::new();
+        let storage = Rc::new(RefCell::new(
+            FileStorage::new(test_file.file_name()).unwrap(),
+        ));
+
+        let storage_index;
+
+        let node1;
+        let node2;
+        let node3;
+
+        let edge1;
+        let edge2;
+        let edge3;
+
+        {
+            let mut graph = DbGraph::new(storage.clone()).unwrap();
+
+            storage_index = graph.storage_index();
+
+            node1 = graph.insert_node().unwrap();
+            node2 = graph.insert_node().unwrap();
+            node3 = graph.insert_node().unwrap();
+
+            edge1 = graph.insert_edge(&node1, &node2).unwrap();
+            edge2 = graph.insert_edge(&node2, &node3).unwrap();
+            edge3 = graph.insert_edge(&node3, &node1).unwrap();
+        }
+
+        let graph = DbGraph::from_storage(storage, storage_index).unwrap();
+
+        assert!(graph.node(&node1).is_some());
+        assert!(graph.node(&node2).is_some());
+        assert!(graph.node(&node3).is_some());
+        assert!(graph.edge(&edge1).is_some());
+        assert!(graph.edge(&edge2).is_some());
+        assert!(graph.edge(&edge3).is_some());
     }
 }
