+++ src/agdb/graph_search/path.rs
@@ -1,4 +1,4 @@
-use crate::graph::graph_index::GraphIndex;
+use crate::graph::GraphIndex;
 
 #[derive(Clone)]
 pub(crate) struct Path {
