+++ src/agdb/graph_search/path_search_handler.rs
@@ -1,4 +1,4 @@
-use crate::graph::graph_index::GraphIndex;
+use crate::graph::GraphIndex;
 
 pub trait PathSearchHandler {
     fn process(&self, index: &GraphIndex, distance: &u64) -> u64;
