+++ src/agdb/graph_search/search_handler.rs
@@ -1,5 +1,5 @@
 use super::search_control::SearchControl;
-use crate::graph::graph_index::GraphIndex;
+use crate::graph::GraphIndex;
 
 pub trait SearchHandler {
     fn process(&self, index: &GraphIndex, distance: &u64) -> SearchControl;
