+++ src/agdb/graph_search/search_iterator.rs
@@ -1,7 +1,7 @@
 use super::search_index::SearchIndex;
-use crate::graph::graph_data::GraphData;
-use crate::graph::graph_impl::GraphImpl;
-use crate::graph::graph_index::GraphIndex;
+use crate::graph::GraphData;
+use crate::graph::GraphImpl;
+use crate::graph::GraphIndex;
 
 pub(crate) trait SearchIterator {
     fn expand_edge<Data: GraphData>(index: &GraphIndex, graph: &GraphImpl<Data>) -> GraphIndex;
